Clean-up the Util class in operator-common
